moved some functions to headers
